MLD_venv/MLD/requirements.txt
dj-database-url==0.5.0
Django==3.1.3
django-crispy-forms==1.9.2
django-haystack==3.0
django-heroku==0.3.1
django-hitcounter==0.1.1
django-revproxy==0.10.0
django-taggit==1.3.0
django-timedeltafield==0.7.10
djangorestframework==3.11.0
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==1.6.0
incremental==17.5.0
numpy==1.17.3
numpydoc==1.1.0
oauth2client==3.0.0
redis==3.5.3
regex==2020.6.8
requests==2.21.0
uritemplate==3.0.1
urllib3==1.24.3
uuid==1.30
virtualenv==20.0.20
waitress==1.4.2